Refatoração criando componente de carrossel e a tela introdutória do app
